requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
aiokafka==0.10.0
redis==5.0.1
boto3==1.34.44
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.0
pydantic-settings==2.5.2
scikit-learn==1.4.2
joblib==1.3.2
